+++ src/server/graphql-server.ts
@@ -1,6 +1,7 @@
 import { ApolloError, ApolloServer, Config, PubSub } from 'apollo-server-koa';
 import { GraphQLSchema } from 'graphql';
 import { applyMiddleware } from 'graphql-middleware';
+import * as http from 'http';
 import { inject, injectable, multiInject } from 'inversify';
 import * as Koa from 'koa';
 import * as koaBody from 'koa-bodyparser';
@@ -13,9 +14,6 @@ import { createMiddleware } from '../middlewares/polaris-middleware-creator';
 import { PolarisProperties } from '../properties/polaris-properties';
 import { PolarisContext } from './polaris-context';
 
-const app = new Koa();
-app.use(koaBody());
-
 export interface GraphQLServer {
     start(): void;
 }
@@ -24,15 +22,17 @@ export type contextCreator = (ctx: Koa.Context) => object;
 
 @injectable()
 export class PolarisGraphQLServer implements GraphQLServer {
-    @inject(POLARIS_TYPES.GraphQLLogger) polarisLogger!: PolarisGraphQLLogger;
+    private app: Koa;
     private server: ApolloServer;
     private polarisProperties: PolarisProperties;
     private customContexts: contextCreator[] = [];
+    private httpServer?: http.Server;
 
     constructor(
         @inject(POLARIS_TYPES.GraphQLSchema) schema: GraphQLSchema,
         @inject(POLARIS_TYPES.PolarisServerConfig) propertiesConfig: PolarisServerConfig,
         @multiInject(POLARIS_TYPES.Middleware) middlewares: Middleware[],
+        @inject(POLARIS_TYPES.GraphQLLogger) private polarisLogger: PolarisGraphQLLogger,
     ) {
         const executableSchemaWithMiddleware = applyMiddleware(
             schema,
@@ -47,16 +47,18 @@ export class PolarisGraphQLServer implements GraphQLServer {
             formatResponse: (response: any) => this.formatResponse(response),
         };
         this.server = new ApolloServer(config);
+        this.app = new Koa();
+        this.app.use(koaBody());
+    }
+
+    start(app = this.app) {
         if (this.polarisProperties.endpoint) {
             this.server.applyMiddleware({ app, path: this.polarisProperties.endpoint });
         } else {
             this.server.applyMiddleware({ app });
         }
-    }
-
-    start() {
         const port = this.polarisProperties.port;
-        const httpServer = app.listen(port, () => {
+        this.httpServer = app.listen(port, () => {
             this.polarisLogger.info(
                 `🚀 Server ready at http://localhost:${port}${this.server.graphqlPath}`,
             );
@@ -65,7 +67,13 @@ export class PolarisGraphQLServer implements GraphQLServer {
             );
         });
         if (this.polarisProperties.includeSubscription) {
-            this.server.installSubscriptionHandlers(httpServer);
+            this.server.installSubscriptionHandlers(this.httpServer);
+        }
+    }
+
+    stop() {
+        if (this.httpServer) {
+            this.httpServer.close();
         }
     }